feat(product-service): upgrade to Java 21, enhance Docker & testing setup
- Dockerfile: switch base image to `openjdk:21-jdk`, package service JAR, run with `java -jar`
- pom.xml: bump project version to `0.0.1`, update Java version to 21
  - add Testcontainers BOM (v1.21.3) for consistent dependency management
  - add dependencies: testcontainers-junit, spring-boot-docker-compose
- Product domain:
  - ProductCreateRequest: switch `price` from double to BigDecimal with validation annotations
  - ProductApplicationService: adjust Money creation to accept BigDecimal directly
- Tests:
  - ProductApplicationServiceTest: add unit test for `getAllProducts`
  - ProductEntityServiceApplicationTests: integrate Testcontainers MongoDB, MockMvc test for product creation
  - ProductJpaRepositoryTest: add repository mapping tests using Mockito

+++ product-service/src/main/java/com/example/product_service/product/application/service/ProductApplicationService.java
@@ -29,7 +29,7 @@ public class ProductApplicationService implements ProductApplicationServiceInter
                 .productId(new ProductId())
                 .name(createRequest.name())
                 .description(createRequest.description())
-                .price(Money.of(BigDecimal.valueOf(createRequest.price()), "EURO"))
+                .price(Money.of(createRequest.price(), "EURO"))
                 .build();
         productRepository.save(product);
     }
